Create maclookup
Sometimes you need to determine the vendor associated with a network MAC address and can't/don't want to go outside the system/container/whatever. So here's a dump of All The Mac Assignments with a built-in way to query them, in one file. Planned mods are obvious: Have a switch to update itself.

Note: I'm not using 'HERE document' syntax because that would defeat the purpose as a self-referential script. :)